create JunitTests for mycoreclass #10
